Major changes to sorting, date format
going to upload, but also then try to merge to master so these are the defaults that will load at regardia.com, not regardia.com/regardia. This is after creating a new folder for atom since using a .js file, and to prepare for a css file.

Still keeping the "WITHSORTING" file to make pushes easier.